+++ src/main/java/jwp/controller/qna/CreateQuestionFormController.java
@@ -1,17 +1,20 @@
-package jwp.controller;
+package jwp.controller.qna;
 
 import core.mvc.Controller;
+import jwp.util.UserSessionUtils;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-public class LogoutController implements Controller {
+public class CreateQuestionFormController implements Controller {
 
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         HttpSession session = req.getSession();
-        session.removeAttribute("user");
-        return "redirect:/";
+        if (UserSessionUtils.isLogined(session)) {          // 회원만 질문 등록 가능
+            return "/qna/form.jsp";
+        }
+        return "redirect:/user/loginForm";
     }
 }
