+++ src/main/java/jwp/controller/qna/ShowQnaController.java
@@ -1,4 +1,4 @@
-package jwp.controller;
+package jwp.controller.qna;
 
 import core.mvc.Controller;
 import jwp.dao.AnswerDao;
@@ -10,7 +10,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
-public class ShowQuestionController implements Controller {
+public class ShowQnaController implements Controller {
 
     private final QuestionDao questionDao = new QuestionDao();
     private final AnswerDao answerDao = new AnswerDao();
