+++ src/main/java/jwp/controller/qna/UpdateQuestionFormController.java
@@ -1,4 +1,4 @@
-package jwp.controller;
+package jwp.controller.qna;
 
 import core.mvc.Controller;
 import jwp.dao.QuestionDao;
@@ -9,7 +9,6 @@ import jwp.util.UserSessionUtils;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
-import java.util.Objects;
 
 
 public class UpdateQuestionFormController implements Controller {
