+++ src/main/java/jwp/controller/user/LogoutController.java
@@ -1,20 +1,17 @@
-package jwp.controller;
+package jwp.controller.user;
 
 import core.mvc.Controller;
-import jwp.util.UserSessionUtils;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-public class CreateQuestionFormController implements Controller {
+public class LogoutController implements Controller {
 
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         HttpSession session = req.getSession();
-        if (UserSessionUtils.isLogined(session)) {          // 회원만 질문 등록 가능
-            return "/qna/form.jsp";
-        }
-        return "redirect:/user/loginForm";
+        session.removeAttribute("user");
+        return "redirect:/";
     }
 }
